+++ course/format/flexsections/classes/output/courseformat/content/section/controlmenu.php
@@ -95,7 +95,7 @@ class controlmenu extends \core_courseformat\output\local\content\section\contro
                     'pixattr' => ['class' => ''],
                     'attr' => [
                         'class' => 'editing_highlight',
-                        'data-action' => 'removemarker'
+                        'data-action' => 'removemarker',
                     ],
                 ];
             } else {
@@ -108,7 +108,7 @@ class controlmenu extends \core_courseformat\output\local\content\section\contro
                     'pixattr' => ['class' => ''],
                     'attr' => [
                         'class' => 'editing_highlight',
-                        'data-action' => 'setmarker'
+                        'data-action' => 'setmarker',
                     ],
                 ];
             }