+++ course/format/flexsections/lang/en/format_flexsections.php
@@ -24,6 +24,7 @@
 
 $string['accordion'] = 'Accordion effect';
 $string['accordiondesc'] = 'When one section is expanded, collapse all others.';
+$string['addsection'] = 'Add section';
 $string['addsections'] = 'Add section';
 $string['addsubsection'] = 'Add subsection';
 $string['backtocourse'] = 'Back to course \'{$a}\'';
@@ -69,8 +70,7 @@ $string['showsection0title_help'] = 'When enabled, the general section will have
 $string['showsection0titledefault'] = 'Show top section title by default';
 $string['showsection0titledefaultdesc'] = 'This defines default setting that will be used for new and existing courses, it can be changed for individual courses in their settings.';
 
-// Deprecated but still can be used in 4.0, to be removed when we have a branch for Moodle 4.1 or later.
-$string['addsection'] = 'Add section';
+// Deprecated since 4.1.
 $string['addsubsectionfor'] = 'Add subsection for \'{$a}\'';
 $string['cancelmoving'] = 'Cancel moving \'{$a}\'';
 $string['removemarker'] = 'Do not mark as current';